move textTemplate to its own type class
