+++ WinampNowPlayingToFile/Business/NowPlayingToFileManager.cs
@@ -11,7 +11,6 @@ using System.Timers;
 using WinampNowPlayingToFile.Facade;
 using WinampNowPlayingToFile.Settings;
 using Song = WinampNowPlayingToFile.Facade.Song;
-using static WinampNowPlayingToFile.Settings.ISettings;
 using System.Drawing.Text;
 
 namespace WinampNowPlayingToFile.Business;
